+++ Source/Alamofire.swift
@@ -36,5 +36,8 @@ import Foundation
 /// Reference to `Session.default` for quick bootstrapping and examples.
 public let AF = Session.default
 
-/// Current Alamofire version. Necessary since SPM doesn't use dynamic libraries. Plus this will be more accurate.
-let version = "5.8.1"
+/// Namespace for informational Alamofire values.
+public enum AFInfo {
+    /// Current Alamofire version.
+    public static let version = "5.8.1"
+}
