+++ Source/Core/Protected.swift
