+++ Source/Core/SessionDelegate.swift
