+++ Source/Core/URLConvertible+URLRequestConvertible.swift
