+++ Source/Extensions/Result+Alamofire.swift
