+++ Source/Extensions/StringEncoding+Alamofire.swift
